Change types for Windows compatibility

+++ swellpy/crepel.pyx
@@ -2,7 +2,7 @@ cimport cython
 import numpy as np
 cimport numpy as np
 
-def iterate (double[:, :] data, long[:] idx, double[:, :, :] kick, long size):
+def iterate (double[:, :] data, long long[:] idx, double[:, :, :] kick, long long size):
     cdef long i, N = size
     for i in range(N):
         data[idx[i], 0] += kick[i, 0, 0]
